+++ generate-example-images/src/main.rs
@@ -52,7 +52,7 @@ fn main() {
     save_jpg("value_cubic", ValueCubic);
 
     save_jpg("fbm", OpenSimplex2.fbm(3, 0.5, 2.0));
-    save_jpg("ridged", OpenSimplex2.ridged(3, 0.5, 2.0));
+    save_jpg("ridged", OpenSimplex2.ridged().fbm(3, 0.5, 2.0));
     save_jpg("ping_pong", OpenSimplex2.ping_pong(3, 0.5, 2.0, 2.0));
 
     save_jpg(